+++ bin/miden-cli/tests/cli.rs
@@ -20,9 +20,7 @@ use miden_client::{
     store::sqlite_store::SqliteStore,
     testing::{
         account_id::ACCOUNT_ID_PRIVATE_SENDER,
-        common::{
-            ACCOUNT_ID_REGULAR, TEST_CLIENT_RPC_CONFIG_FILE, execute_tx_and_sync, insert_new_wallet,
-        },
+        common::{ACCOUNT_ID_REGULAR, TEST_CLIENT_RPC_CONFIG_FILE},
     },
     transaction::{OutputNote, TransactionRequestBuilder},
     utils::Serializable,
@@ -452,7 +450,8 @@ async fn debug_mode_outputs_logs() {
     // Create a Client and a custom note
     let store_path = create_test_store_path();
     let (mut client, authenticator) = create_rust_client_with_store_path(&store_path).await;
-    let (account, ..) = insert_new_wallet(&mut client, AccountStorageMode::Private, &authenticator)
+    let (account, ..) = client
+        .insert_new_wallet(AccountStorageMode::Private, &authenticator)
         .await
         .unwrap();
 
@@ -484,7 +483,7 @@ async fn debug_mode_outputs_logs() {
         .own_output_notes(vec![OutputNote::Full(note.clone())])
         .build()
         .unwrap();
-    execute_tx_and_sync(&mut client, account.id(), transaction_request).await;
+    client.execute_tx_and_sync(account.id(), transaction_request).await.unwrap();
 
     // Export the note
     let note_file: NoteFile = NoteFile::NoteDetails {