+++ tests/src/main_tests.rs
@@ -15,7 +15,7 @@ use miden_objects::{
     account::AccountStorageMode,
     asset::{Asset, FungibleAsset},
     note::{NoteFile, NoteType},
-    transaction::{ProvenTransaction, ToInputNoteCommitments, TransactionWitness},
+    transaction::{ProvenTransaction, TransactionWitness},
 };
 use winter_maybe_async::maybe_async_trait;
 
@@ -64,16 +64,20 @@ async fn multiple_tx_on_same_block() {
     let (mut client, authenticator) = create_test_client().await;
     wait_for_node(&mut client).await;
 
-    let (first_regular_account, second_regular_account, faucet_account_header) =
-        setup_two_wallets_and_faucet(&mut client, AccountStorageMode::Private, &authenticator)
-            .await;
+    let (wallets, faucets) = client
+        .setup_accounts_and_faucets(
+            &authenticator,
+            AccountStorageMode::Private,
+            2,
+            1,
+            vec![vec![MINT_AMOUNT, 0]],
+        )
+        .await
+        .unwrap();
 
-    let from_account_id = first_regular_account.id();
-    let to_account_id = second_regular_account.id();
-    let faucet_account_id = faucet_account_header.id();
-
-    // First Mint necessary token
-    mint_and_consume(&mut client, from_account_id, faucet_account_id, NoteType::Private).await;
+    let from_account_id = wallets[0].id();
+    let to_account_id = wallets[1].id();
+    let faucet_account_id = faucets[0].id();
 
     // Do a transfer from first account to second account
     let asset = FungibleAsset::new(faucet_account_id, TRANSFER_AMOUNT).unwrap();
@@ -120,7 +124,7 @@ async fn multiple_tx_on_same_block() {
     client.sync_state().await.unwrap();
 
     // wait for 1 block
-    wait_for_blocks(&mut client, 1).await;
+    client.wait_for_blocks(1).await.unwrap();
 
     // Submit the proven transactions
     client.testing_submit_proven_transaction(tx_prove_1).await.unwrap();
@@ -159,10 +163,10 @@ async fn import_expected_notes() {
         setup_wallet_and_faucet(&mut client_1, AccountStorageMode::Private, &authenticator_1).await;
 
     let (mut client_2, authenticator_2) = create_test_client().await;
-    let (client_2_account, _seed, _) =
-        insert_new_wallet(&mut client_2, AccountStorageMode::Private, &authenticator_2)
-            .await
-            .unwrap();
+    let (client_2_account, _seed, _) = client_2
+        .insert_new_wallet(AccountStorageMode::Private, &authenticator_2)
+        .await
+        .unwrap();
 
     wait_for_node(&mut client_2).await;
 
@@ -180,10 +184,10 @@ async fn import_expected_notes() {
 
     // If the verification is requested before execution then the import should fail
     assert!(client_2.import_note(NoteFile::NoteId(note.id())).await.is_err());
-    execute_tx_and_sync(&mut client_1, faucet_account.id(), tx_request).await;
+    client_1.execute_tx_and_sync(faucet_account.id(), tx_request).await.unwrap();
 
     // Use client 1 to wait until a couple of blocks have passed
-    wait_for_blocks(&mut client_1, 3).await;
+    client_1.wait_for_blocks(3).await.unwrap();
 
     let new_sync_data = client_2.sync_state().await.unwrap();
 
@@ -199,7 +203,7 @@ async fn import_expected_notes() {
     assert!(input_note.inclusion_proof().is_some());
 
     // If client 2 successfully consumes the note, we confirm we have MMR and block header data
-    consume_notes(&mut client_2, client_2_account.id(), &[input_note.try_into().unwrap()]).await;
+    consume_notes(&mut client_2, client_2_account.id(), &[input_note]).await;
 
     let tx_request = TransactionRequestBuilder::new()
         .build_mint_fungible_asset(
@@ -227,7 +231,7 @@ async fn import_expected_notes() {
     // If imported before execution then the inclusion proof should be None
     assert!(input_note.inclusion_proof().is_none());
 
-    execute_tx_and_sync(&mut client_1, faucet_account.id(), tx_request).await;
+    client_1.execute_tx_and_sync(faucet_account.id(), tx_request).await.unwrap();
     client_2.sync_state().await.unwrap();
 
     // After sync, the imported note should have inclusion proof even if it's not relevant for its
@@ -236,23 +240,23 @@ async fn import_expected_notes() {
     assert!(input_note.inclusion_proof().is_some());
 
     // If inclusion proof is invalid this should panic
-    consume_notes(&mut client_1, first_basic_account.id(), &[input_note.try_into().unwrap()]).await;
+    consume_notes(&mut client_1, first_basic_account.id(), &[input_note]).await;
 }
 
 #[tokio::test]
 async fn import_expected_note_uncommitted() {
     let (mut client_1, authenticator) = create_test_client().await;
-    let faucet_account =
-        insert_new_fungible_faucet(&mut client_1, AccountStorageMode::Private, &authenticator)
-            .await
-            .unwrap()
-            .0;
+    let faucet_account = client_1
+        .insert_new_fungible_faucet(AccountStorageMode::Private, &authenticator)
+        .await
+        .unwrap()
+        .0;
 
     let (mut client_2, _) = create_test_client().await;
-    let (client_2_account, _seed, _) =
-        insert_new_wallet(&mut client_2, AccountStorageMode::Private, &authenticator)
-            .await
-            .unwrap();
+    let (client_2_account, _seed, _) = client_2
+        .insert_new_wallet(AccountStorageMode::Private, &authenticator)
+        .await
+        .unwrap();
 
     wait_for_node(&mut client_2).await;
 
@@ -307,10 +311,10 @@ async fn import_expected_notes_from_the_past_as_committed() {
 
     let block_height_before = client_1.get_sync_height().await.unwrap();
 
-    execute_tx_and_sync(&mut client_1, faucet_account.id(), tx_request).await;
+    client_1.execute_tx_and_sync(faucet_account.id(), tx_request).await.unwrap();
 
     // Use client 1 to wait until a couple of blocks have passed
-    wait_for_blocks(&mut client_1, 3).await;
+    client_1.wait_for_blocks(3).await.unwrap();
     client_2.sync_state().await.unwrap();
 
     // If the verification is requested before execution then the import should fail
@@ -340,14 +344,12 @@ async fn get_account_update() {
         setup_wallet_and_faucet(&mut client, AccountStorageMode::Private, &authenticator).await;
     wait_for_node(&mut client).await;
 
-    let (basic_wallet_2, ..) =
-        insert_new_wallet(&mut client, AccountStorageMode::Public, &authenticator)
-            .await
-            .unwrap();
+    let (basic_wallet_2, ..) = client
+        .insert_new_wallet(AccountStorageMode::Public, &authenticator)
+        .await
+        .unwrap();
 
-    // Mint and consume notes with both accounts so they are included in the node.
-    mint_and_consume(&mut client, basic_wallet_1.id(), faucet_account.id(), NoteType::Private)
-        .await;
+    // Mint and consume notes with the second account to deploy it.
     mint_and_consume(&mut client, basic_wallet_2.id(), faucet_account.id(), NoteType::Private)
         .await;
 
@@ -373,10 +375,10 @@ async fn sync_detail_values() {
     let (first_regular_account, faucet_account_header) =
         setup_wallet_and_faucet(&mut client1, AccountStorageMode::Private, &authenticator_1).await;
 
-    let (second_regular_account, ..) =
-        insert_new_wallet(&mut client2, AccountStorageMode::Private, &authenticator_2)
-            .await
-            .unwrap();
+    let (second_regular_account, ..) = client2
+        .insert_new_wallet(AccountStorageMode::Private, &authenticator_2)
+        .await
+        .unwrap();
 
     let from_account_id = first_regular_account.id();
     let to_account_id = second_regular_account.id();
@@ -404,7 +406,7 @@ async fn sync_detail_values() {
         )
         .unwrap();
     let note_id = tx_request.expected_output_own_notes().pop().unwrap().id();
-    execute_tx_and_sync(&mut client1, from_account_id, tx_request).await;
+    client1.execute_tx_and_sync(from_account_id, tx_request).await.unwrap();
 
     // Second client sync should have new note
     let new_details = client2.sync_state().await.unwrap();
@@ -415,7 +417,7 @@ async fn sync_detail_values() {
 
     // Consume the note with the second account
     let tx_request = TransactionRequestBuilder::new().build_consume_notes(vec![note_id]).unwrap();
-    execute_tx_and_sync(&mut client2, to_account_id, tx_request).await;
+    client2.execute_tx_and_sync(to_account_id, tx_request).await.unwrap();
 
     // First client sync should have a new nullifier as the note was consumed
     let new_details = client1.sync_state().await.unwrap();
@@ -581,16 +583,25 @@ async fn consume_multiple_expected_notes() {
     wait_for_node(&mut client).await;
 
     // Setup accounts
-    let (target_basic_account_1, faucet_account_header) =
-        setup_wallet_and_faucet(&mut client, AccountStorageMode::Private, &authenticator_1).await;
-    let (target_basic_account_2, ..) =
-        insert_new_wallet(&mut unauth_client, AccountStorageMode::Private, &authenticator_2)
-            .await
-            .unwrap();
+    let (wallets, faucets) = client
+        .setup_accounts_and_faucets(
+            &authenticator_1,
+            AccountStorageMode::Private,
+            1,
+            1,
+            vec![vec![0]],
+        )
+        .await
+        .unwrap();
+
+    let (target_basic_account_2, ..) = unauth_client
+        .insert_new_wallet(AccountStorageMode::Private, &authenticator_2)
+        .await
+        .unwrap();
     unauth_client.sync_state().await.unwrap();
 
-    let faucet_account_id = faucet_account_header.id();
-    let to_account_ids = [target_basic_account_1.id(), target_basic_account_2.id()];
+    let faucet_account_id = faucets[0].id();
+    let to_account_ids = [wallets[0].id(), target_basic_account_2.id()];
 
     // Mint tokens to the accounts
     let fungible_asset = FungibleAsset::new(faucet_account_id, TRANSFER_AMOUNT).unwrap();
@@ -601,7 +612,10 @@ async fn consume_multiple_expected_notes() {
         client.rng(),
     );
 
-    execute_tx_and_sync(&mut client, faucet_account_id, mint_tx_request.clone()).await;
+    client
+        .execute_tx_and_sync(faucet_account_id, mint_tx_request.clone())
+        .await
+        .unwrap();
     unauth_client.sync_state().await.unwrap();
 
     // Filter notes by ownership
@@ -623,8 +637,8 @@ async fn consume_multiple_expected_notes() {
         .build()
         .unwrap();
 
-    let tx_id_1 = execute_tx(&mut client, to_account_ids[0], tx_request_1).await;
-    let tx_id_2 = execute_tx(&mut unauth_client, to_account_ids[1], tx_request_2).await;
+    let tx_id_1 = client.execute_tx(to_account_ids[0], tx_request_1).await.unwrap();
+    let tx_id_2 = unauth_client.execute_tx(to_account_ids[1], tx_request_2).await.unwrap();
 
     // Ensure notes are processed
     assert!(!client.get_input_notes(NoteFilter::Processing).await.unwrap().is_empty());
@@ -667,10 +681,10 @@ async fn import_consumed_note_with_proof() {
         setup_wallet_and_faucet(&mut client_1, AccountStorageMode::Private, &authenticator_1).await;
 
     let (mut client_2, authenticator_2) = create_test_client().await;
-    let (client_2_account, _seed, _) =
-        insert_new_wallet(&mut client_2, AccountStorageMode::Private, &authenticator_2)
-            .await
-            .unwrap();
+    let (client_2_account, _seed, _) = client_2
+        .insert_new_wallet(AccountStorageMode::Private, &authenticator_2)
+        .await
+        .unwrap();
 
     wait_for_node(&mut client_2).await;
 
@@ -696,7 +710,7 @@ async fn import_consumed_note_with_proof() {
             client_1.rng(),
         )
         .unwrap();
-    execute_tx_and_sync(&mut client_1, from_account_id, tx_request).await;
+    client_1.execute_tx_and_sync(from_account_id, tx_request).await.unwrap();
     let note = client_1
         .get_input_notes(NoteFilter::Committed)
         .await
@@ -709,7 +723,7 @@ async fn import_consumed_note_with_proof() {
 
     println!("Consuming Note...");
     let tx_request = TransactionRequestBuilder::new().build_consume_notes(vec![note.id()]).unwrap();
-    execute_tx_and_sync(&mut client_1, from_account_id, tx_request).await;
+    client_1.execute_tx_and_sync(from_account_id, tx_request).await.unwrap();
 
     // Import the consumed note
     client_2
@@ -727,19 +741,25 @@ async fn import_consumed_note_with_proof() {
 #[tokio::test]
 async fn import_consumed_note_with_id() {
     let (mut client_1, authenticator) = create_test_client().await;
-    let (first_regular_account, second_regular_account, faucet_account_header) =
-        setup_two_wallets_and_faucet(&mut client_1, AccountStorageMode::Private, &authenticator)
-            .await;
+
+    wait_for_node(&mut client_1).await;
+
+    let (wallets, faucets) = client_1
+        .setup_accounts_and_faucets(
+            &authenticator,
+            AccountStorageMode::Private,
+            2,
+            1,
+            vec![vec![MINT_AMOUNT, 0]],
+        )
+        .await
+        .unwrap();
+
+    let from_account_id = wallets[0].id();
+    let to_account_id = wallets[1].id();
+    let faucet_account_id = faucets[0].id();
 
     let (mut client_2, _) = create_test_client().await;
-
-    wait_for_node(&mut client_2).await;
-
-    let from_account_id = first_regular_account.id();
-    let to_account_id = second_regular_account.id();
-    let faucet_account_id = faucet_account_header.id();
-
-    mint_and_consume(&mut client_1, from_account_id, faucet_account_id, NoteType::Private).await;
 
     let current_block_num = client_1.get_sync_height().await.unwrap();
     let asset = FungibleAsset::new(faucet_account_id, TRANSFER_AMOUNT).unwrap();
@@ -757,7 +777,7 @@ async fn import_consumed_note_with_id() {
             client_1.rng(),
         )
         .unwrap();
-    execute_tx_and_sync(&mut client_1, from_account_id, tx_request).await;
+    client_1.execute_tx_and_sync(from_account_id, tx_request).await.unwrap();
     let note = client_1
         .get_input_notes(NoteFilter::Committed)
         .await
@@ -770,7 +790,7 @@ async fn import_consumed_note_with_id() {
 
     println!("Consuming Note...");
     let tx_request = TransactionRequestBuilder::new().build_consume_notes(vec![note.id()]).unwrap();
-    execute_tx_and_sync(&mut client_1, from_account_id, tx_request).await;
+    client_1.execute_tx_and_sync(from_account_id, tx_request).await.unwrap();
     client_2.sync_state().await.unwrap();
 
     // Import the consumed note
@@ -783,22 +803,25 @@ async fn import_consumed_note_with_id() {
 #[tokio::test]
 async fn import_note_with_proof() {
     let (mut client_1, authenticator) = create_test_client().await;
-    let (first_regular_account, second_regular_account, faucet_account_header) =
-        setup_two_wallets_and_faucet(&mut client_1, AccountStorageMode::Private, &authenticator)
-            .await;
+
+    wait_for_node(&mut client_1).await;
+
+    let (wallets, faucets) = client_1
+        .setup_accounts_and_faucets(
+            &authenticator,
+            AccountStorageMode::Private,
+            2,
+            1,
+            vec![vec![MINT_AMOUNT, 0]],
+        )
+        .await
+        .unwrap();
+
+    let from_account_id = wallets[0].id();
+    let to_account_id = wallets[1].id();
+    let faucet_account_id = faucets[0].id();
 
     let (mut client_2, _) = create_test_client().await;
-
-    wait_for_node(&mut client_2).await;
-
-    let from_account_id = first_regular_account.id();
-    let to_account_id = second_regular_account.id();
-    let faucet_account_id = faucet_account_header.id();
-
-    let note =
-        mint_note(&mut client_1, from_account_id, faucet_account_id, NoteType::Private).await;
-
-    consume_notes(&mut client_1, from_account_id, &[note]).await;
 
     let current_block_num = client_1.get_sync_height().await.unwrap();
     let asset = FungibleAsset::new(faucet_account_id, TRANSFER_AMOUNT).unwrap();
@@ -816,7 +839,7 @@ async fn import_note_with_proof() {
             client_1.rng(),
         )
         .unwrap();
-    execute_tx_and_sync(&mut client_1, from_account_id, tx_request).await;
+    client_1.execute_tx_and_sync(from_account_id, tx_request).await.unwrap();
 
     let note = client_1
         .get_input_notes(NoteFilter::Committed)
@@ -850,10 +873,10 @@ async fn discarded_transaction() {
         setup_wallet_and_faucet(&mut client_1, AccountStorageMode::Private, &authenticator_1).await;
 
     let (mut client_2, authenticator_2) = create_test_client().await;
-    let (second_regular_account, ..) =
-        insert_new_wallet(&mut client_2, AccountStorageMode::Private, &authenticator_2)
-            .await
-            .unwrap();
+    let (second_regular_account, ..) = client_2
+        .insert_new_wallet(AccountStorageMode::Private, &authenticator_2)
+        .await
+        .unwrap();
 
     wait_for_node(&mut client_2).await;
 
@@ -880,7 +903,7 @@ async fn discarded_transaction() {
         )
         .unwrap();
 
-    execute_tx_and_sync(&mut client_1, from_account_id, tx_request).await;
+    client_1.execute_tx_and_sync(from_account_id, tx_request).await.unwrap();
     client_2.sync_state().await.unwrap();
     let note = client_1
         .get_input_notes(NoteFilter::Committed)
@@ -918,7 +941,7 @@ async fn discarded_transaction() {
     assert!(matches!(note_record.state(), InputNoteState::ProcessingAuthenticated(_)));
 
     // Consume the note in client 2
-    execute_tx_and_sync(&mut client_2, to_account_id, tx_request).await;
+    client_2.execute_tx_and_sync(to_account_id, tx_request).await.unwrap();
 
     let note_record = client_2.get_input_note(note.id()).await.unwrap().unwrap();
     assert!(matches!(note_record.state(), InputNoteState::ConsumedAuthenticatedLocal(_)));
@@ -1010,15 +1033,15 @@ async fn custom_transaction_prover() {
 async fn locked_account() {
     let (mut client_1, authenticator) = create_test_client().await;
 
-    let (faucet_account, ..) =
-        insert_new_fungible_faucet(&mut client_1, AccountStorageMode::Private, &authenticator)
-            .await
-            .unwrap();
+    let (faucet_account, ..) = client_1
+        .insert_new_fungible_faucet(AccountStorageMode::Private, &authenticator)
+        .await
+        .unwrap();
 
-    let (private_account, seed, _) =
-        insert_new_wallet(&mut client_1, AccountStorageMode::Private, &authenticator)
-            .await
-            .unwrap();
+    let (private_account, seed, _) = client_1
+        .insert_new_wallet(AccountStorageMode::Private, &authenticator)
+        .await
+        .unwrap();
 
     let from_account_id = private_account.id();
     let faucet_account_id = faucet_account.id();
@@ -1062,15 +1085,15 @@ async fn locked_account() {
 #[tokio::test]
 async fn expired_transaction_fails() {
     let (mut client, authenticator) = create_test_client().await;
-    let (faucet_account, ..) =
-        insert_new_fungible_faucet(&mut client, AccountStorageMode::Private, &authenticator)
-            .await
-            .unwrap();
+    let (faucet_account, ..) = client
+        .insert_new_fungible_faucet(AccountStorageMode::Private, &authenticator)
+        .await
+        .unwrap();
 
-    let (private_account, ..) =
-        insert_new_wallet(&mut client, AccountStorageMode::Private, &authenticator)
-            .await
-            .unwrap();
+    let (private_account, ..) = client
+        .insert_new_wallet(AccountStorageMode::Private, &authenticator)
+        .await
+        .unwrap();
 
     let from_account_id = private_account.id();
     let faucet_account_id = faucet_account.id();
@@ -1092,7 +1115,7 @@ async fn expired_transaction_fails() {
         client.new_transaction(faucet_account_id, tx_request).await.unwrap();
 
     println!("Transaction executed successfully");
-    wait_for_blocks(&mut client, (expiration_delta + 1).into()).await;
+    client.wait_for_blocks((expiration_delta + 1).into()).await.unwrap();
 
     println!("Sending transaction to node");
     let submitted_tx_result = client.submit_transaction(transaction_execution_result).await;
@@ -1166,13 +1189,22 @@ async fn unused_rpc_api() {
 #[tokio::test]
 async fn ignore_invalid_notes() {
     let (mut client, authenticator) = create_test_client().await;
-    let (regular_account, second_regular_account, faucet_account_header) =
-        setup_two_wallets_and_faucet(&mut client, AccountStorageMode::Private, &authenticator)
-            .await;
+    wait_for_node(&mut client).await;
 
-    let account_id = regular_account.id();
-    let second_account_id = second_regular_account.id();
-    let faucet_account_id = faucet_account_header.id();
+    let (wallets, faucets) = client
+        .setup_accounts_and_faucets(
+            &authenticator,
+            AccountStorageMode::Private,
+            2,
+            1,
+            vec![vec![0, 0]],
+        )
+        .await
+        .unwrap();
+
+    let account_id = wallets[0].id();
+    let second_account_id = wallets[1].id();
+    let faucet_account_id = faucets[0].id();
 
     // Mint 2 valid notes
     let note_1 = mint_note(&mut client, account_id, faucet_account_id, NoteType::Private).await;
@@ -1190,7 +1222,7 @@ async fn ignore_invalid_notes() {
         .build_consume_notes(vec![note_1.id(), note_3.id(), note_2.id(), note_4.id()])
         .unwrap();
 
-    execute_tx_and_sync(&mut client, account_id, tx_request).await;
+    client.execute_tx_and_sync(account_id, tx_request).await.unwrap();
 
     // Check that only the valid notes were consumed
     let consumed_notes = client.get_input_notes(NoteFilter::Consumed).await.unwrap();